tests/wpt/web-platform-tests/app-history/navigate/return-value/goTo-cross-document-preventDefault.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  let key = i.contentWindow.appHistory.current.key;

  i.contentWindow.appHistory.navigate("?1");
  await new Promise(resolve => i.onload = resolve);

  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[1]);

  // This will be a noop, because navigate events are uncancelable for traversals.
  i.contentWindow.appHistory.onnavigate = e => e.preventDefault();

  assertNeverSettles(t, i.contentWindow.appHistory.goTo(key), i.contentWindow);
  await new Promise(resolve => i.onload = () => t.step_timeout(resolve, 0));
}, "goTo() promise never settle when preventDefault()ing the navigate event (cross-document)");
</script>
